cloud_function/requirements.txt
google-cloud-storage==2.8.0
google-cloud-pubsub==2.15.0
pandas==1.5.3
great_expectations==0.16.16
numpy==1.25.0
